added draggable element

+++ Photobox.Web/Photobox.Web.Client/Pages/Components/ImageObject.cs
@@ -3,8 +3,8 @@
 public class ImageObject
 {
     public string Content { get; set; } = ""; // Image URL or Text
-    public int X { get; set; } = 0;
-    public int Y { get; set; } = 0;
+    public double X { get; set; } = 0;
+    public double Y { get; set; } = 0;
     
     public double Scale { get; set; } = 1.0d;
 }
